数据库/实验/期末实验/Final_project/back-end/pyproject.toml
[project]
name = "sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12.2"
dependencies = [
    "mcp[cli]>=1.3.0",
    "mysqlclient>=2.2.7",
    "httpx>=0.25.0",
    "python-dotenv>=1.0.0",
]
